+++ library/build.gradle
@@ -28,6 +28,10 @@ android {
         disable 'InvalidPackage'
     }
 
+    testOptions {
+        unitTests.returnDefaultValues = true
+    }
+
 }
 
 def supportVersion = '23.4.0'
